requirements.txt
uv>=0.1.0
python-dotenv>=1.0.0
httpx>=0.25.0 